util/playlist-extractor/requirements.txt
spotipy==2.23.0
pandas==2.1.4
python-dotenv==1.0.0
pymongo==4.6.1
requests==2.31.0 